Product_recognition/requirements.txt
matplotlib
numpy
scikit-learn
torch
torchvision
shutil
Pillow
opencv-python
pandas
os
tqdm
itertools

